feat: 결제 기능 추가(근데 결제 정보 저장이나 확인은 안되고 취소도 안되는)

+++ src/main/java/com/bookdream/sbb/pay/PayRepository.java
@@ -2,6 +2,6 @@ package com.bookdream.sbb.pay;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PayRepository  extends JpaRepository<Pay, Long> {
+public interface PayRepository  extends JpaRepository<Pay, Integer> {
 
 }
